services/liveness-engine/requirements.txt
fastapi
uvicorn
numpy
scikit-image
pillow
opencv-python-headless
